feat: add logic to generate rss files
