+++ .gitignore
@@ -67,8 +67,6 @@ public/sitemap-0.xml
 sitemap.xml
 sitemap-0.xml
 
-rss.xml
-
 # Mac files
 .DS_Store
