+++ package.json
@@ -66,6 +66,7 @@
     "react-use": "^17.4.0",
     "remark-gfm": "^3.0.1",
     "rive-react": "^3.0.34",
+    "rss": "^1.2.2",
     "search-insights": "^2.2.3",
     "slugify": "^1.6.5",
     "swagger-ui": "^4.15.5",
@@ -86,7 +87,6 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.32.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "feed": "^4.2.2",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.0",
     "postcss": "^8.4.21",
